Give access to the underlying SqlCommand
Think of this as an 'escape hatch', for when a user wants to do something we haven't thought about. Many TPs do something similar.

I have one use case for this: Be able to change the CommandText at run-time. In a reporting application I made long ago, there's an Sql template, which is then manipulated for constraining the results in ways that are just not possible using @params.
In this Type Provider, the Sql template would provide the type, but at run-time I could use the modified query instead.